refactor: Replace ConfigFileSettings.CreateGlobal
with GitConfigSettings

Refs: #12201
