+++ src/app/GitUI/CommandsDialogs/SettingsDialog/CheckSettingsLogic.cs
@@ -2,7 +2,6 @@
 using GitCommands.Git;
 using GitCommands.Utils;
 using GitExtensions.Extensibility.Git;
-using GitUI.CommandsDialogs.SettingsDialog.Pages;
 using Microsoft.Win32;
 
 namespace GitUI.CommandsDialogs.SettingsDialog
@@ -41,7 +40,7 @@ namespace GitUI.CommandsDialogs.SettingsDialog
 
             if (string.IsNullOrEmpty(editor))
             {
-                Environment.SetEnvironmentVariable(CommonLogic.AmbientGitEditorEnvVariableName, EditorHelper.FileEditorCommand);
+                Environment.SetEnvironmentVariable(CommonLogic.AmbientGitEditorEnvVariableName, AppSettings.FileEditorCommand);
             }
 
             return true;
